fix broken retry

+++ internal/config/config.go
@@ -40,7 +40,7 @@ type APIKeyConfig struct {
 
 type NatsJetstreamConfig struct {
 	URL             string                   `mapstructure:"url"`
-	MaxRetries      int                      `mapstructure:"max_retries"`
+	MaxRetries      uint64                   `mapstructure:"max_retries"`
 	ReconnectFactor float64                  `mapstructure:"reconnect_factor"`
 	MinJitter       time.Duration            `mapstructure:"min_jitter"`
 	MaxJitter       time.Duration            `mapstructure:"max_jitter"`
